re-use alarm discard modal popup content
