+++ frontend/src/modules/controllers/ModalPopup.tsx
@@ -18,6 +18,14 @@ import {
   IconButton,
 } from '@material-ui/core';
 import ReplyIcon from '@material-ui/icons/Reply';
+import { useSelector } from 'react-redux';
+import { AlarmConfiguration, alarmConfiguration } from '../alarms/AlarmsPage';
+import {
+  getAlarmLimitsRequestDraft,
+  getAlarmLimitsRequest,
+  getAlarmLimitsUnsavedKeys,
+} from '../../store/controller/selectors';
+import { VentilationMode, Range } from '../../store/controller/proto/mcu_pb';
 
 const useStyles = makeStyles((theme: Theme) => ({
   closeButton: {
@@ -49,6 +57,16 @@ const useStyles = makeStyles((theme: Theme) => ({
         },
       },
     },
+  },
+  marginContent: {
+    textAlign: 'center',
+    marginTop: theme.spacing(2),
+    marginBottom: theme.spacing(3),
+  },
+  marginHeader: {
+    textAlign: 'center',
+    marginTop: theme.spacing(3),
+    marginBottom: theme.spacing(1),
   },
   containerClass: {
     flexDirection: 'column',
@@ -114,6 +132,7 @@ interface Props {
   showCloseIcon?: boolean;
   onClose?(): void;
   onConfirm?(): void;
+  discardContent?: boolean;
   maxWidth?: DialogProps['maxWidth'];
 }
 
@@ -163,6 +182,46 @@ const ModalAction = ({ onClose, onConfirm }: ActionProps): JSX.Element => {
   );
 };
 
+const AlarmDiscardContent = (): JSX.Element => {
+  const classes = useStyles();
+  const alarmLimitsRequestDraftSelect = useSelector(getAlarmLimitsRequestDraft);
+  const alarmLimitsRequestSelect = useSelector(getAlarmLimitsRequest);
+  const alarmLimitsUnsavedKeys = useSelector(getAlarmLimitsUnsavedKeys);
+  const alarmLimitsRequest = (alarmLimitsRequestSelect as unknown) as Record<string, Range>;
+  const alarmLimitsRequestDraft = (alarmLimitsRequestDraftSelect as unknown) as Record<
+    string,
+    Range
+  >;
+  const alarmConfig = alarmConfiguration(VentilationMode.hfnc);
+
+  return (
+    <Grid container alignItems="center">
+      <Grid container alignItems="center" justify="center">
+        <Grid container alignItems="center" className={classes.marginHeader}>
+          <Grid item xs>
+            <Typography variant="h4">Keep Previous Values?</Typography>
+          </Grid>
+        </Grid>
+        <Grid item className={classes.marginContent}>
+          {alarmConfig.map((param: AlarmConfiguration) => {
+            if (alarmLimitsRequest !== null && alarmLimitsRequestDraft !== null) {
+              if (alarmLimitsUnsavedKeys.includes(param.stateKey)) {
+                return (
+                  <Typography variant="subtitle1">{`Keep ${param.label} alarm range to ${
+                    alarmLimitsRequest[param.stateKey].lower
+                  } -
+                          ${alarmLimitsRequest[param.stateKey].upper}?`}</Typography>
+                );
+              }
+            }
+            return <React.Fragment />;
+          })}
+        </Grid>
+      </Grid>
+    </Grid>
+  );
+};
+
 /**
  * ModalPopup
  *
@@ -181,6 +240,7 @@ export const ModalPopup = (props: PropsWithChildren<Props>): JSX.Element => {
     showCloseIcon,
     children,
     withAction,
+    discardContent,
     onConfirm,
     fullWidth = true,
     maxWidth = 'xl',
@@ -214,7 +274,7 @@ export const ModalPopup = (props: PropsWithChildren<Props>): JSX.Element => {
           </DialogTitle>
           <Grid container alignItems="center" className={classes.containerClass}>
             <DialogContent>
-              {children}
+              {discardContent ? <AlarmDiscardContent></AlarmDiscardContent> : children}
               <DialogActions style={{ width: '100%' }} className={classes.dialogActions}>
                 {withAction && <ModalAction onClose={onClose} onConfirm={onConfirm} />}
               </DialogActions>
